requirements.txt
bs4
MechanicalSoup
requests
pandas
flask
plotly
openpyxl
dash_bootstrap_components
dash_bootstrap_templates
gunicorn
dash
FlightRadarAPI
bokeh
pygal
matplotlib
seaborn
numpy
opencv-python
pillow
cairosvg
typer